Move updates code from xdl to cli
