+++ packages/expo-cli/package.json
@@ -101,11 +101,14 @@
     "js-yaml": "^3.13.1",
     "keychain": "1.3.0",
     "klaw-sync": "6.0.0",
+    "latest-version": "5.1.0",
     "leven": "^3.1.0",
     "lodash": "4.17.15",
     "md5-file": "^5.0.0",
     "npm-package-arg": "6.1.0",
     "ora": "3.4.0",
+    "p-timeout": "3.1.0",
+    "package-json": "6.4.0",
     "pacote": "^11.1.0",
     "pngjs": "3.4.0",
     "progress": "2.0.3",
