+++ packages/xdl/package.json
@@ -66,7 +66,6 @@
     "internal-ip": "4.3.0",
     "invariant": "2.2.4",
     "is-reachable": "^4.0.0",
-    "latest-version": "5.1.0",
     "lodash": "4.17.15",
     "md5hex": "1.0.0",
     "minimatch": "3.0.4",
@@ -77,7 +76,6 @@
     "p-map": "3.0.0",
     "p-retry": "4.1.0",
     "p-timeout": "3.1.0",
-    "package-json": "6.4.0",
     "pacote": "11.1.0",
     "pascal-case": "2.0.1",
     "pretty-bytes": "^5.3.0",
